backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
pydantic-settings
python-dotenv
tencentcloud-sdk-python-lighthouse
tortoise-orm
